Move am43 sensor code and remove auto load on cover
