+++ esphome/components/am43/__init__.py
@@ -1,0 +1,1 @@
+CODEOWNERS = ["@buxtronix"]
