+++ esphome/components/am43/sensor/am43_sensor.cpp
@@ -1,6 +1,6 @@
-#include "am43.h"
-#include "esphome/core/log.h"
+#include "am43_sensor.h"
 #include "esphome/core/hal.h"
+#include "esphome/core/log.h"
 
 #ifdef USE_ESP32
